+++ public/app/features/dashboard-scene/scene/layout-default/row-actions/RowActions.tsx
@@ -1,31 +1,18 @@
-import { css } from '@emotion/css';
-
-import { GrafanaTheme2 } from '@grafana/data';
-import {
-  SceneComponentProps,
-  sceneGraph,
-  SceneGridRow,
-  SceneObjectBase,
-  SceneObjectState,
-  VizPanel,
-} from '@grafana/scenes';
-import { Icon, TextLink, useStyles2 } from '@grafana/ui';
+import { sceneGraph, SceneGridRow, SceneObjectBase, SceneObjectState } from '@grafana/scenes';
 import appEvents from 'app/core/app_events';
-import { t, Trans } from 'app/core/internationalization';
-import { SHARED_DASHBOARD_QUERY } from 'app/plugins/datasource/dashboard/constants';
-import { MIXED_DATASOURCE_NAME } from 'app/plugins/datasource/mixed/MixedDataSource';
+import { t } from 'app/core/internationalization';
 import { ShowConfirmModalEvent } from 'app/types/events';
 
-import { getDashboardSceneFor, getQueryRunnerFor } from '../../../utils/utils';
-import { DashboardGridItem } from '../DashboardGridItem';
 import { DefaultGridLayoutManager } from '../DefaultGridLayoutManager';
 import { RowRepeaterBehavior } from '../RowRepeaterBehavior';
 
-import { RowOptionsButton } from './RowOptionsButton';
+import { RowActionsRenderer } from './RowActionsRenderer';
 
 export interface RowActionsState extends SceneObjectState {}
 
 export class RowActions extends SceneObjectBase<RowActionsState> {
+  static Component = RowActionsRenderer;
+
   public getParent(): SceneGridRow {
     if (!(this.parent instanceof SceneGridRow)) {
       throw new Error('RowActions must have a SceneGridRow parent');
@@ -39,7 +26,7 @@ export class RowActions extends SceneObjectBase<RowActionsState> {
     manager.removeRow(this.getParent(), removePanels);
   }
 
-  public onUpdate = (title: string, repeat?: string | null): void => {
+  public onUpdate(title: string, repeat: string | null | undefined) {
     const row = this.getParent();
     let repeatBehavior: RowRepeaterBehavior | undefined;
 
@@ -56,8 +43,7 @@ export class RowActions extends SceneObjectBase<RowActionsState> {
     }
 
     if (repeat) {
-      // Remove repeat behavior if it exists
-      // to retrigger repeat when adding new one
+      // Remove repeat behavior if it exists to re-trigger repeat when adding new one
       if (repeatBehavior) {
         repeatBehavior.removeBehavior();
       }
@@ -67,9 +53,9 @@ export class RowActions extends SceneObjectBase<RowActionsState> {
     } else {
       repeatBehavior?.removeBehavior();
     }
-  };
+  }
 
-  public onDelete = () => {
+  public onDelete() {
     appEvents.publish(
       new ShowConfirmModalEvent({
         title: t('dashboard.default-layout.row-actions.modal.title', 'Delete row'),
@@ -83,106 +69,5 @@ export class RowActions extends SceneObjectBase<RowActionsState> {
         onAltAction: () => this.removeRow(),
       })
     );
-  };
-
-  public getWarning = () => {
-    const row = this.getParent();
-    const gridItems = row.state.children;
-
-    const isAnyPanelUsingDashboardDS = gridItems.some((gridItem) => {
-      if (!(gridItem instanceof DashboardGridItem)) {
-        return false;
-      }
-
-      const vizPanel = gridItem.state.body;
-      if (vizPanel instanceof VizPanel) {
-        const runner = getQueryRunnerFor(vizPanel);
-        return (
-          runner?.state.datasource?.uid === SHARED_DASHBOARD_QUERY ||
-          (runner?.state.datasource?.uid === MIXED_DATASOURCE_NAME &&
-            runner?.state.queries.some((query) => query.datasource?.uid === SHARED_DASHBOARD_QUERY))
-        );
-      }
-
-      return false;
-    });
-
-    if (isAnyPanelUsingDashboardDS) {
-      return (
-        <div>
-          <p>
-            <Trans i18nKey="dashboard.default-layout.row-actions.repeat.warning.text">
-              Panels in this row use the {{ SHARED_DASHBOARD_QUERY }} data source. These panels will reference the panel
-              in the original row, not the ones in the repeated rows.
-            </Trans>
-          </p>
-          <TextLink
-            external
-            href={
-              'https://grafana.com/docs/grafana/latest/dashboards/build-dashboards/create-dashboard/#configure-repeating-rows'
-            }
-          >
-            <Trans i18nKey="dashboard.default-layout.row-actions.repeat.warning.learn-more">Learn more</Trans>
-          </TextLink>
-        </div>
-      );
-    }
-
-    return undefined;
-  };
-
-  static Component = ({ model }: SceneComponentProps<RowActions>) => {
-    const dashboard = getDashboardSceneFor(model);
-    const row = model.getParent();
-    const { title } = row.useState();
-    const { meta, isEditing } = dashboard.useState();
-    const styles = useStyles2(getStyles);
-
-    const behaviour = row.state.$behaviors?.find((b) => b instanceof RowRepeaterBehavior);
-
-    return (
-      <>
-        {meta.canEdit && isEditing && (
-          <>
-            <div className={styles.rowActions}>
-              <RowOptionsButton
-                title={title}
-                repeat={behaviour instanceof RowRepeaterBehavior ? behaviour.state.variableName : undefined}
-                parent={dashboard}
-                onUpdate={model.onUpdate}
-                warning={model.getWarning()}
-              />
-              <button
-                type="button"
-                onClick={model.onDelete}
-                aria-label={t('dashboard.default-layout.row-actions.delete', 'Delete row')}
-              >
-                <Icon name="trash-alt" />
-              </button>
-            </div>
-          </>
-        )}
-      </>
-    );
-  };
+  }
 }
-
-const getStyles = (theme: GrafanaTheme2) => {
-  return {
-    rowActions: css({
-      color: theme.colors.text.secondary,
-      lineHeight: '27px',
-
-      button: {
-        color: theme.colors.text.secondary,
-        paddingLeft: theme.spacing(2),
-        background: 'transparent',
-        border: 'none',
-
-        '&:hover': {
-          color: theme.colors.text.maxContrast,
-        },
-      },
-    }),
-  };
-};
