+++ public/app/features/dashboard-scene/scene/layout-default/row-actions/RowOptionsButton.tsx
@@ -1,5 +1,3 @@
-import { ReactNode } from 'react';
-
 import { SceneObject } from '@grafana/scenes';
 import { Icon, ModalsController } from '@grafana/ui';
 import { t } from 'app/core/internationalization';
@@ -12,15 +10,10 @@ export interface RowOptionsButtonProps {
   repeat?: string;
   parent: SceneObject;
   onUpdate: OnRowOptionsUpdate;
-  warning?: ReactNode;
+  isUsingDashboardDS: boolean;
 }
 
-export const RowOptionsButton = ({ repeat, title, parent, onUpdate, warning }: RowOptionsButtonProps) => {
-  const onUpdateChange = (hideModal: () => void) => (title: string, repeat?: string | null) => {
-    onUpdate(title, repeat);
-    hideModal();
-  };
-
+export const RowOptionsButton = ({ repeat, title, parent, onUpdate, isUsingDashboardDS }: RowOptionsButtonProps) => {
   return (
     <ModalsController>
       {({ showModal, hideModal }) => {
@@ -35,8 +28,11 @@ export const RowOptionsButton = ({ repeat, title, parent, onUpdate, warning }: R
                 repeat,
                 parent,
                 onDismiss: hideModal,
-                onUpdate: onUpdateChange(hideModal),
-                warning,
+                onUpdate: (title: string, repeat?: string | null) => {
+                  onUpdate(title, repeat);
+                  hideModal();
+                },
+                isUsingDashboardDS,
               });
             }}
           >
