+++ public/app/features/dashboard-scene/scene/layout-default/row-actions/RowOptionsForm.tsx
@@ -1,11 +1,12 @@
-import { ReactNode, useCallback, useState } from 'react';
+import { useCallback, useState } from 'react';
 import { useForm } from 'react-hook-form';
 
 import { selectors } from '@grafana/e2e-selectors';
 import { SceneObject } from '@grafana/scenes';
-import { Button, Field, Modal, Input, Alert } from '@grafana/ui';
+import { Button, Field, Modal, Input, Alert, TextLink } from '@grafana/ui';
 import { t, Trans } from 'app/core/internationalization';
 import { RepeatRowSelect2 } from 'app/features/dashboard/components/RepeatRowSelect/RepeatRowSelect';
+import { SHARED_DASHBOARD_QUERY } from 'app/plugins/datasource/dashboard/constants';
 
 export type OnRowOptionsUpdate = (title: string, repeat?: string | null) => void;
 
@@ -15,17 +16,15 @@ export interface Props {
   sceneContext: SceneObject;
   onUpdate: OnRowOptionsUpdate;
   onCancel: () => void;
-  warning?: ReactNode;
+  isUsingDashboardDS: boolean;
 }
 
-export const RowOptionsForm = ({ repeat, title, sceneContext, warning, onUpdate, onCancel }: Props) => {
+export const RowOptionsForm = ({ repeat, title, sceneContext, isUsingDashboardDS, onUpdate, onCancel }: Props) => {
   const [newRepeat, setNewRepeat] = useState<string | undefined>(repeat);
   const onChangeRepeat = useCallback((name?: string) => setNewRepeat(name), [setNewRepeat]);
 
   const { handleSubmit, register } = useForm({
-    defaultValues: {
-      title,
-    },
+    defaultValues: { title },
   });
 
   const submit = (formData: { title: string }) => {
@@ -40,7 +39,7 @@ export const RowOptionsForm = ({ repeat, title, sceneContext, warning, onUpdate,
       <Field label={t('dashboard.default-layout.row-options.form.repeat-for', 'Repeat for')}>
         <RepeatRowSelect2 sceneContext={sceneContext} repeat={newRepeat} onChange={onChangeRepeat} />
       </Field>
-      {warning && (
+      {isUsingDashboardDS && (
         <Alert
           data-testid={selectors.pages.Dashboard.Rows.Repeated.ConfigSection.warningMessage}
           severity="warning"
@@ -48,7 +47,22 @@ export const RowOptionsForm = ({ repeat, title, sceneContext, warning, onUpdate,
           topSpacing={3}
           bottomSpacing={0}
         >
-          {warning}
+          <div>
+            <p>
+              <Trans i18nKey="dashboard.default-layout.row-actions.repeat.warning.text">
+                Panels in this row use the {{ SHARED_DASHBOARD_QUERY }} data source. These panels will reference the
+                panel in the original row, not the ones in the repeated rows.
+              </Trans>
+            </p>
+            <TextLink
+              external
+              href={
+                'https://grafana.com/docs/grafana/latest/dashboards/build-dashboards/create-dashboard/#configure-repeating-rows'
+              }
+            >
+              <Trans i18nKey="dashboard.default-layout.row-actions.repeat.warning.learn-more">Learn more</Trans>
+            </TextLink>
+          </div>
         </Alert>
       )}
       <Modal.ButtonRow>
